Minor fixes.
- Fixed cast to size_t instead of uint32_t.
- Style fixes.

+++ Pcap++/src/DpdkDeviceList.cpp
@@ -204,7 +204,7 @@ DpdkDevice* DpdkDeviceList::getDeviceByPort(int portId) const
 		return nullptr;
 	}
 
-	if (static_cast<uint32_t>(portId) >= m_DeviceList.size())
+	if (static_cast<std::size_t>(portId) >= m_DeviceList.size())
 	{
 		return nullptr;
 	}
